+++ src/index.d.ts
@@ -1,7 +1,17 @@
 import { VNode } from 'preact';
 
-export default function renderToString(vnode: VNode, context?: any): string;
+export default function renderToString<P = {}>(
+	vnode: VNode<P>,
+	context?: any
+): string;
 
-export function render(vnode: VNode, context?: any): string;
-export function renderToString(vnode: VNode, context?: any): string;
-export function renderToStaticMarkup(vnode: VNode, context?: any): string;
+export function render<P = {}>(vnode: VNode<P>, context?: any): string;
+export function renderToString<P = {}>(vnode: VNode<P>, context?: any): string;
+export function renderToStringAsync<P = {}>(
+	vnode: VNode<P>,
+	context?: any
+): string | Promise<string>;
+export function renderToStaticMarkup<P = {}>(
+	vnode: VNode<P>,
+	context?: any
+): string;